+++ messaging-core/src/main/java/com/nabla/sdk/messaging/core/domain/entity/MessageInput.kt
@@ -1,0 +1,20 @@
+package com.nabla.sdk.messaging.core.domain.entity
+
+import com.nabla.sdk.core.domain.entity.FileUpload
+
+public sealed class MessageInput {
+    /**
+     * Input for a text message, containing only the text to send
+     */
+    public data class Text(val text: String) : MessageInput()
+    public sealed class Media : MessageInput() {
+        /**
+         * Input for an image message, containing the local image to send
+         */
+        public data class Image(val mediaSource: FileSource.Local<FileLocal.Image, FileUpload.Image>) : Media()
+        /**
+         * Input for an document message, containing the local document to send
+         */
+        public data class Document(val mediaSource: FileSource.Local<FileLocal.Document, FileUpload.Document>) : Media()
+    }
+}