+++ messaging-core/src/test/java/com/nabla/sdk/messaging/core/data/stubs/NablaMessagingClientStub.kt
@@ -7,8 +7,8 @@ import com.nabla.sdk.messaging.core.NablaMessagingClient
 import com.nabla.sdk.messaging.core.domain.entity.Conversation
 import com.nabla.sdk.messaging.core.domain.entity.ConversationId
 import com.nabla.sdk.messaging.core.domain.entity.ConversationItems
-import com.nabla.sdk.messaging.core.domain.entity.Message
 import com.nabla.sdk.messaging.core.domain.entity.MessageId
+import com.nabla.sdk.messaging.core.domain.entity.MessageInput
 import com.nabla.sdk.messaging.core.domain.entity.WatchPaginatedResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -70,9 +70,12 @@ class NablaMessagingClientStub(
             }
     }
 
-    override suspend fun sendMessage(message: Message): Result<Unit> {
+    override suspend fun sendMessage(
+        input: MessageInput,
+        conversationId: ConversationId
+    ): Result<MessageId.Local> {
         return runCatchingCancellable {
-            conversationContentRepository.sendMessage(message)
+            conversationContentRepository.sendMessage(input, conversationId)
         }
     }